+++ sdks/full/rust/src/models/cloud_games_validate_game_request.rs
@@ -16,16 +16,16 @@ pub struct CloudGamesValidateGameRequest {
     /// Represent a resource's readable display name.
     #[serde(rename = "display_name")]
     pub display_name: String,
-    /// A human readable short identifier used to references resources. Different than a `rivet.common#Uuid` because this is intended to be human readable. Different than `rivet.common#DisplayName` because this should not include special characters and be short.
-    #[serde(rename = "name_id")]
-    pub name_id: String,
+    /// **Deprecated** A human readable short identifier used to references resources. Different than a `rivet.common#Uuid` because this is intended to be human readable. Different than `rivet.common#DisplayName` because this should not include special characters and be short.
+    #[serde(rename = "name_id", skip_serializing_if = "Option::is_none")]
+    pub name_id: Option<String>,
 }
 
 impl CloudGamesValidateGameRequest {
-    pub fn new(display_name: String, name_id: String) -> CloudGamesValidateGameRequest {
+    pub fn new(display_name: String) -> CloudGamesValidateGameRequest {
         CloudGamesValidateGameRequest {
             display_name,
-            name_id,
+            name_id: None,
         }
     }
 }