+++ sdks/runtime/typescript/src/api/resources/matchmaker/resources/lobbies/client/Client.ts
@@ -27,6 +27,7 @@ export class Lobbies {
 
     /**
      * Marks the current lobby as ready to accept connections. Players will not be able to connect to this lobby until the lobby is flagged as ready.
+     * This endpoint requires a [lobby token](/docs/general/concepts/token-types#matchmaker-lobby) for authentication, or a [development namespace token](/docs/general/concepts/token-types#namespace-development) for mock responses. When running on Rivet servers, you can access the given lobby token from the [`RIVET_TOKEN`](/docs/matchmaker/concepts/lobby-env) environment variable.
      * @throws {@link Rivet.InternalError}
      * @throws {@link Rivet.RateLimitError}
      * @throws {@link Rivet.ForbiddenError}
@@ -142,6 +143,11 @@ export class Lobbies {
      * join using the /join endpoint (this can be disabled by the developer by rejecting all new connections
      * after setting the lobby to closed).
      * Does not shutdown the lobby.
+     *
+     * This endpoint requires a [lobby token](/docs/general/concepts/token-types#matchmaker-lobby) for
+     * authentication, or a [development namespace token](/docs/general/concepts/token-types#namespace-development)
+     * for mock responses. When running on Rivet servers, you can access the given lobby token from the
+     * [`RIVET_TOKEN`](/docs/matchmaker/concepts/lobby-env) environment variable.
      * @throws {@link Rivet.InternalError}
      * @throws {@link Rivet.RateLimitError}
      * @throws {@link Rivet.ForbiddenError}
@@ -259,6 +265,12 @@ export class Lobbies {
     }
 
     /**
+     * Sets the state JSON of the current lobby.
+     *
+     * This endpoint requires a [lobby token](/docs/general/concepts/token-types#matchmaker-lobby) for
+     * authentication, or a [development namespace token](/docs/general/concepts/token-types#namespace-development)
+     * for mock responses. When running on Rivet servers, you can access the given lobby token from the
+     * [`RIVET_TOKEN`](/docs/matchmaker/concepts/lobby-env) environment variable.
      * @throws {@link Rivet.InternalError}
      * @throws {@link Rivet.RateLimitError}
      * @throws {@link Rivet.ForbiddenError}
@@ -376,6 +388,12 @@ export class Lobbies {
     }
 
     /**
+     * Get the state of any lobby.
+     *
+     * This endpoint requires a [lobby token](/docs/general/concepts/token-types#matchmaker-lobby) for
+     * authentication, or a [development namespace token](/docs/general/concepts/token-types#namespace-development)
+     * for mock responses. When running on Rivet servers, you can access the given lobby token from the
+     * [`RIVET_TOKEN`](/docs/matchmaker/concepts/lobby-env) environment variable.
      * @throws {@link Rivet.InternalError}
      * @throws {@link Rivet.RateLimitError}
      * @throws {@link Rivet.ForbiddenError}
@@ -496,6 +514,12 @@ export class Lobbies {
      * Finds a lobby based on the given criteria.
      * If a lobby is not found and `prevent_auto_create_lobby` is `false`,
      * a new lobby will be created.
+     *
+     * When [tokenless authentication](/docs/general/concepts/tokenless-authentication/web) is enabled in
+     * your game namespace, this endpoint does not require a token to authenticate. Otherwise, a
+     * [development namespace token](/docs/general/concepts/token-types#namespace-development) can be used
+     * for mock responses and a [public namespace token](/docs/general/concepts/token-types#namespace-public)
+     * can be used for general authentication.
      * @throws {@link Rivet.InternalError}
      * @throws {@link Rivet.RateLimitError}
      * @throws {@link Rivet.ForbiddenError}
@@ -622,6 +646,12 @@ export class Lobbies {
      * Joins a specific lobby.
      * This request will use the direct player count configured for the
      * lobby group.
+     *
+     * When [tokenless authentication](/docs/general/concepts/tokenless-authentication/web) is enabled in
+     * your game namespace, this endpoint does not require a token to authenticate. Otherwise, a
+     * [development namespace token](/docs/general/concepts/token-types#namespace-development) can be used
+     * for mock responses and a [public namespace token](/docs/general/concepts/token-types#namespace-public)
+     * can be used for general authentication.
      * @throws {@link Rivet.InternalError}
      * @throws {@link Rivet.RateLimitError}
      * @throws {@link Rivet.ForbiddenError}
@@ -746,6 +776,12 @@ export class Lobbies {
 
     /**
      * Creates a custom lobby.
+     *
+     * When [tokenless authentication](/docs/general/concepts/tokenless-authentication/web) is enabled in
+     * your game namespace, this endpoint does not require a token to authenticate. Otherwise, a
+     * [development namespace token](/docs/general/concepts/token-types#namespace-development) can be used
+     * for mock responses and a [public namespace token](/docs/general/concepts/token-types#namespace-public)
+     * can be used for general authentication.
      * @throws {@link Rivet.InternalError}
      * @throws {@link Rivet.RateLimitError}
      * @throws {@link Rivet.ForbiddenError}
@@ -870,6 +906,12 @@ export class Lobbies {
 
     /**
      * Lists all open lobbies.
+     *
+     * When [tokenless authentication](/docs/general/concepts/tokenless-authentication/web) is enabled in
+     * your game namespace, this endpoint does not require a token to authenticate. Otherwise, a
+     * [development namespace token](/docs/general/concepts/token-types#namespace-development) can be used
+     * for mock responses and a [public namespace token](/docs/general/concepts/token-types#namespace-public)
+     * can be used for general authentication.
      * @throws {@link Rivet.InternalError}
      * @throws {@link Rivet.RateLimitError}
      * @throws {@link Rivet.ForbiddenError}